move default contenttypes to "contenttypes" package.
In order to cleanup up extras dependencies conditions, we move all
things related to default content types into the subpackage
ftw.simplelayout.contenttypes so that we can only include this package
when the corresponding extras was installed.
This makes it easier to not break Plone 4 installations without the
contenttypes extras installed, because we have a clear separation for
lib things and contenttype things.
