+++ ftw/simplelayout/contenttypes/browser/galleryblock.py
