+++ ftw/simplelayout/contenttypes/browser/templates/galleryblock.pt
