+++ ftw/simplelayout/contenttypes/browser/templates/listingblock.pt
