+++ ftw/simplelayout/contenttypes/browser/templates/textblock.pt
