+++ ftw/simplelayout/contenttypes/browser/textblock.py
@@ -1,7 +1,7 @@
 from Acquisition._Acquisition import aq_inner
-from ftw.simplelayout.behaviors import ITeaser
 from ftw.simplelayout.browser.blocks.base import BaseBlock
-from ftw.simplelayout.contents.textblock import ITextBlockSchema
+from ftw.simplelayout.contenttypes.behaviors import ITeaser
+from ftw.simplelayout.contenttypes.contents.textblock import ITextBlockSchema
 from ftw.simplelayout.interfaces import ISimplelayoutActions
 from ftw.simplelayout.utils import normalize_portal_type
 from plone.memoize.instance import memoize
