+++ ftw/simplelayout/contenttypes/browser/uploadcapable.py
