+++ ftw/simplelayout/contenttypes/browser/videoblock.py
@@ -1,6 +1,6 @@
 from ftw.simplelayout.browser.blocks.base import BaseBlock
-from ftw.simplelayout.contents.videoblock import is_vimeo_url
-from ftw.simplelayout.contents.videoblock import is_youtube_url
+from ftw.simplelayout.contenttypes.contents.videoblock import is_vimeo_url
+from ftw.simplelayout.contenttypes.contents.videoblock import is_youtube_url
 from plone.uuid.interfaces import IUUID
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from urlparse import urlparse
