+++ ftw/simplelayout/contenttypes/configure.zcml
@@ -1,0 +1,31 @@
+<configure
+    xmlns="http://namespaces.zope.org/zope"
+    xmlns:genericsetup="http://namespaces.zope.org/genericsetup"
+    xmlns:i18n="http://namespaces.zope.org/i18n"
+    xmlns:plone="http://namespaces.plone.org/plone"
+    xmlns:zcml="http://namespaces.zope.org/zcml"
+    i18n_domain="ftw.simplelayout">
+
+    <include package="plone.behavior" file="meta.zcml" />
+
+    <include file="lawgiver.zcml" zcml:condition="installed ftw.lawgiver" />
+    <include file="permissions.zcml" />
+    <include package=".browser" />
+    <include package=".contents" />
+
+    <plone:behavior
+        title="Simplelayout teaser behavior"
+        description="Extends a content by two fields internal and external link"
+        provides="ftw.simplelayout.contenttypes.behaviors.ITeaser"
+        factory="plone.behavior.AnnotationStorage"
+        />
+
+    <genericsetup:registerProfile
+        name="default"
+        title="ftw.simplelayout.contenttypes default"
+        directory="profiles/default"
+        description="Simlelayout default content types"
+        provides="Products.GenericSetup.interfaces.EXTENSION"
+        />
+
+</configure>
