+++ ftw/simplelayout/contenttypes/contents/contentpage.py
@@ -1,7 +1,7 @@
+from ftw.simplelayout.contenttypes.contents.interfaces import IContentPage
 from plone.dexterity.content import Container
 from plone.supermodel import model
 from zope.interface import implements
-from ftw.simplelayout.contents.interfaces import IContentPage
 
 
 class IContentPageSchema(model.Schema):
