+++ ftw/simplelayout/contenttypes/contents/filelistingblock.py
@@ -1,8 +1,8 @@
 from collections import OrderedDict
 from ftw.simplelayout import _
 from ftw.simplelayout.browser.actions import DefaultActions
-from ftw.simplelayout.contents.interfaces import IFileListingBlock
-from ftw.simplelayout.contents.interfaces import IListingBlockColumns
+from ftw.simplelayout.contenttypes.contents.interfaces import IFileListingBlock
+from ftw.simplelayout.contenttypes.contents.interfaces import IListingBlockColumns
 from ftw.table import helper
 from plone.autoform.interfaces import IFormFieldProvider
 from plone.dexterity.content import Container
