+++ ftw/simplelayout/contenttypes/contents/galleryblock.py
@@ -1,7 +1,7 @@
 from collections import OrderedDict
 from ftw.simplelayout import _
 from ftw.simplelayout.browser.actions import DefaultActions
-from ftw.simplelayout.contents.interfaces import IGalleryBlock
+from ftw.simplelayout.contenttypes.contents.interfaces import IGalleryBlock
 from plone.autoform.interfaces import IFormFieldProvider
 from plone.dexterity.content import Container
 from plone.directives import form
