+++ ftw/simplelayout/contenttypes/contents/interfaces.py
