+++ ftw/simplelayout/contenttypes/profiles/default/jsregistry.xml
