+++ ftw/simplelayout/contenttypes/profiles/default/metadata.xml
