+++ ftw/simplelayout/contenttypes/profiles/default/propertiestool.xml
