+++ ftw/simplelayout/contenttypes/profiles/default/rolemap.xml
@@ -21,11 +21,6 @@
       <role name="Site Administrator"/>
       <role name="Contributor" />
     </permission>
-    <permission name="ftw.simplelayout: Change Layouts" acquire="False">
-      <role name="Manager" />
-      <role name="Site Administrator"/>
-      <role name="Contributor" />
-    </permission>
     <permission name="ftw.simplelayout: Add GalleryBlock" acquire="False">
       <role name="Manager" />
       <role name="Site Administrator"/>
