+++ ftw/simplelayout/contenttypes/profiles/default/tinymce.xml
