+++ ftw/simplelayout/contenttypes/profiles/default/types.xml
