+++ ftw/simplelayout/contenttypes/profiles/default/types/ftw.simplelayout.VideoBlock.xml
@@ -1,28 +1,26 @@
 <?xml version="1.0"?>
-<object name="ftw.simplelayout.FileListingBlock"
+<object name="ftw.simplelayout.VideoBlock"
         meta_type="Dexterity FTI"
         xmlns:i18n="http://xml.zope.org/namespaces/i18n"
         i18n:domain="ftw.simplelayout">
 
     <!-- Basic metadata -->
-    <property name="title" i18n:translate="">FileListingBlock</property>
-    <property name="description" i18n:translate="">The file listing block renders a list of uploaded files with configurable header which can be used to change the order of the listing.</property>
+    <property name="title" i18n:translate="">VideoBlock</property>
+    <property name="description" i18n:translate="">The video block renders videos loaded from YouTube or Vimeo.</property>
     <property name="icon_expr"></property>
     <property name="allow_discussion">False</property>
     <property name="global_allow">False</property>
     <property name="filter_content_types">True</property>
-    <property name="allowed_content_types">
-        <element value="File" />
-    </property>
+    <property name="allowed_content_types" />
 
     <!-- schema interface -->
-    <property name="schema">ftw.simplelayout.contents.filelistingblock.IFileListingBlockSchema</property>
+    <property name="schema">ftw.simplelayout.contenttypes.contents.videoblock.IVideoBlockSchema</property>
 
     <!-- class used for content items -->
-    <property name="klass">ftw.simplelayout.contents.filelistingblock.FileListingBlock</property>
+    <property name="klass">ftw.simplelayout.contenttypes.contents.videoblock.VideoBlock</property>
 
     <!-- add permission -->
-    <property name="add_permission">ftw.simplelayout.AddFileListingBlock</property>
+    <property name="add_permission">ftw.simplelayout.AddVideoBlock</property>
 
     <!-- enabled behaviors -->
     <property name="behaviors">
@@ -32,10 +30,10 @@
     </property>
 
     <!-- View information -->
-    <property name="default_view">redirect_to_parent</property>
+    <property name="default_view">@@redirect_to_parent</property>
     <property name="default_view_fallback">False</property>
     <property name="view_methods">
-        <element value="redirect_to_parent"/>
+        <element value="@@redirect_to_parent"/>
     </property>
 
     <!-- Method aliases -->
