+++ ftw/simplelayout/contenttypes/profiles/default/workflows.xml
