+++ ftw/simplelayout/lawgiver.zcml
@@ -6,17 +6,8 @@
     <include package="ftw.lawgiver" file="meta.zcml" />
 
     <lawgiver:map_permissions
-        action_group="add"
-        permissions="ftw.simplelayout: Add ContentPage"
-        />
-
-    <lawgiver:map_permissions
         action_group="edit"
-        permissions="ftw.simplelayout: Add FileListingBlock,
-                     ftw.simplelayout: Add TextBlock,
-                     ftw.simplelayout: Add VideoBlock,
-                     ftw.simplelayout: Change Layouts,
-                     ftw.simplelayout: Add GalleryBlock"
+        permissions="ftw.simplelayout: Change Layouts"
         />
 
 </configure>
