+++ ftw/simplelayout/permissions.zcml
@@ -3,33 +3,8 @@
     i18n_domain="ftw.simplelayout">
 
     <permission
-        id="ftw.simplelayout.AddContentPage"
-        title="ftw.simplelayout: Add ContentPage"
-        />
-
-    <permission
-        id="ftw.simplelayout.AddTextBLock"
-        title="ftw.simplelayout: Add TextBlock"
-        />
-
-    <permission
-        id="ftw.simplelayout.AddFileListingBlock"
-        title="ftw.simplelayout: Add FileListingBlock"
-        />
-
-    <permission
-        id="ftw.simplelayout.AddVideoBlock"
-        title="ftw.simplelayout: Add VideoBlock"
-        />
-
-    <permission
         id="ftw.simplelayout.ChangeLayouts"
         title="ftw.simplelayout: Change Layouts"
-        />
-
-    <permission
-        id="ftw.simplelayout.AddGalleryBlock"
-        title="ftw.simplelayout: Add GalleryBlock"
         />
 
 </configure>
