+++ ftw/simplelayout/profiles/lib-plone5/rolemap.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+<rolemap>
+  <permissions>
+    <permission name="ftw.simplelayout: Change Layouts" acquire="False">
+      <role name="Manager" />
+      <role name="Site Administrator"/>
+      <role name="Contributor" />
+    </permission>
+  </permissions>
+</rolemap>
