+++ ftw/simplelayout/tests/test_uploadcapable.py
@@ -1,7 +1,7 @@
 from collective.quickupload.interfaces import IQuickUploadFileFactory
 from ftw.builder import Builder
 from ftw.builder import create
-from ftw.simplelayout.browser import uploadcapable
+from ftw.simplelayout.contenttypes.browser import uploadcapable
 from ftw.simplelayout.testing import FTW_SIMPLELAYOUT_CONTENT_TESTING
 from unittest2 import TestCase
 from zope.component import queryAdapter
